Changed type of xy spline parameters to be List[Float] instead of List[Int].

+++ q2_ps_plot/plugin_setup.py
@@ -45,8 +45,8 @@ shared_parameters = {
     "color_by": Str,
     "flex_reps": Bool,
     "psea": Bool,
-    "spline_x": List[Int],
-    "spline_y": List[Int]
+    "spline_x": List[Float],
+    "spline_y": List[Float]
 }
 # shared parameter decriptions for zenrich and zenrich tsv pipeline
 shared_descriptions = {
